Use new format for dynamic localization bindings
