+++ src/NexusMods.App.UI/Localization/Localizer.cs
@@ -2,7 +2,10 @@ using System.ComponentModel;
 using System.Globalization;
 using System.Text;
 using Avalonia;
+using Avalonia.Data;
+using Avalonia.Data.Core;
 using Avalonia.Markup.Xaml.MarkupExtensions;
+using Avalonia.Markup.Xaml.MarkupExtensions.CompiledBindings;
 using NexusMods.App.UI.Resources;
 
 namespace NexusMods.App.UI.Localization;
@@ -55,5 +58,60 @@ public class Localizer : INotifyPropertyChanged // <= INotifyPropertyChanged is 
     {
         PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(IndexerName));
         PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(IndexerArrayName));
+    }
+
+    internal static CompiledBindingExtension GetBindingForKey(string key)
+    {
+        // Tip: This binds the [] method of the Localizer class.
+        // Build binding for '$"[{Key}]"'.
+        var x = new CompiledBindingPathBuilder();
+        x = x.SetRawSource(Localizer.Instance);
+        x = x.Property(
+            new ClrPropertyInfo(
+                "Item",
+                obj => ((Localizer)obj)[key],
+                null,
+                typeof(string)),
+            PropertyInfoAccessorFactory.CreateInpcPropertyAccessor);
+
+        var binding = new CompiledBindingExtension(x.Build())
+        {
+            Mode = BindingMode.OneWay,
+            Source = Localizer.Instance,
+        };
+        return binding;
+    }
+
+
+
+    /// <summary>
+    /// Creates a dynamic binding for a localized string.
+    /// The binding will update when the locale is changed.
+    /// The new localized value will be retrieved using the provided function.
+    /// </summary>
+    /// <example>
+    /// <code>
+    /// [!TextBlock.TextProperty] = Localizer.CreateBinding(() => Language.Helpers_GenerateHeader_NAME)
+    /// </code>
+    /// </example>
+    /// <param name="getLocalizedString">A function that returns a string localized for the current locale.</param>
+    public static CompiledBindingExtension CreateBinding(Func<string> getLocalizedString)
+    {
+        var x = new CompiledBindingPathBuilder();
+        x = x.SetRawSource(Localizer.Instance);
+        x = x.Property(
+            new ClrPropertyInfo(
+                "Item",
+                obj => getLocalizedString(),
+                null,
+                typeof(string)),
+            PropertyInfoAccessorFactory.CreateInpcPropertyAccessor);
+
+        var binding = new CompiledBindingExtension(x.Build())
+        {
+            Mode = BindingMode.OneWay,
+            Source = Localizer.Instance,
+        };
+        return binding;
     }
 }
